requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow-cpu==2.15.0
numpy==1.24.3
scikit-learn==1.3.0
pydantic==2.5.0